fix: route file-size drift to issues
