+++ ui-dashboard/src/lib/network-fetcher/constants.ts
@@ -1,1 +1,2 @@
 export const SNAPSHOT_PAGE_SIZE = 1000;
+export const SECONDS_PER_DAY = 86_400;
