+++ ui-dashboard/src/lib/network-fetcher/fetch.ts
@@ -2,13 +2,11 @@
 // (`@/hooks/use-all-networks-data`) and Server Components that need to
 // pre-render the dashboard payload. Deliberately avoids any React / SWR
 // imports so it can run in both runtimes without tripping Next.js's RSC
-// swr/react-server bundling (which has no default export). Module-scope
-// state (`warnedCapKeys`, `partialPageLastCapturedAt`) intentionally
-// persists across requests to throttle Sentry signals across the 30s poll
-// loop; tests `.clear()` these to reset between cases.
+// swr/react-server bundling (which has no default export). Pagination module
+// state intentionally persists across requests to throttle Sentry signals
+// across the 30s poll loop; tests `.clear()` those caches between cases.
 
 import { GraphQLClient } from "graphql-request";
-import * as Sentry from "@sentry/nextjs";
 import { NETWORKS, NETWORK_IDS, isConfiguredNetworkId } from "@/lib/networks";
 import type { Network } from "@/lib/networks";
 import {
@@ -21,10 +19,6 @@ import {
   ALL_POOLS_WITH_HEALTH,
   ALL_OLS_POOLS,
   ALL_CDP_POOLS,
-  BROKER_DAILY_SNAPSHOTS_ALL,
-  POOL_DAILY_FEE_SNAPSHOTS_PAGE,
-  POOL_DAILY_SNAPSHOTS_ALL,
-  UNIQUE_LP_ADDRESSES,
 } from "@/lib/queries";
 import { aggregateProtocolFees } from "@/lib/protocol-fees";
 import { activeReservePoolIdsFromKnownStrategies } from "@/lib/strategy-contracts";
@@ -36,18 +30,11 @@ import {
   type SnapshotWindows,
   type TimeRange,
 } from "@/lib/volume";
-import type {
-  Pool,
-  PoolDailyFeeSnapshot,
-  PoolSnapshotWindow,
-  CdpPool,
-} from "@/lib/types";
+import type { Pool, CdpPool } from "@/lib/types";
 import { isFpmm, buildOracleRateMap } from "@/lib/tokens";
 import type {
-  BrokerDailySnapshotRow,
   NetworkData,
   OlsPoolsResult,
-  PaginatedPageResult,
   PoolBreachRollupResult,
   PoolHealthCursorResult,
   PoolRebalanceThresholdsKnownResult,
@@ -56,13 +43,32 @@ import type {
   SnapshotPageResult,
   VpLifecycleDeprecationResult,
 } from "./types";
-import { SNAPSHOT_PAGE_SIZE } from "./constants";
 import {
   mergeDeprecatedVirtualPools,
   type VpExchangeDeprecationResult,
 } from "./vp-deprecation";
+import {
+  REQUEST_TIMEOUT_MS,
+  fetchAllBrokerDailySnapshotPages,
+  fetchAllDailySnapshotPages,
+  fetchAllFeeSnapshotPages,
+  fetchAllLpAddressPages,
+  mutableDayCutoff,
+} from "./pagination";
+
+import { SECONDS_PER_DAY } from "./constants";
 
 export { SNAPSHOT_PAGE_SIZE } from "./constants";
+export {
+  REQUEST_TIMEOUT_MS,
+  fetchAllDailySnapshotPages,
+  fetchAllFeeSnapshotPages,
+  fetchPaginatedRows,
+  incrementalRowCache,
+  partialPageLastCapturedAt,
+  seedIncrementalRowCacheFromNetworkData,
+  warnedCapKeys,
+} from "./pagination";
 
 /**
  * True iff every error channel on `n` is null — top-level, rates,
@@ -143,549 +149,6 @@ const emptyNetworkData = (
   snapshotWindows: SnapshotWindows,
   error: Error,
 ): NetworkData => blankNetworkData(network, snapshotWindows, { error });
-
-/**
- * Envio's hosted Hasura silently caps every query at 1000 rows regardless of
- * the `limit` we send, so for full history we paginate with offset. Stop as
- * soon as a page comes back under the page size (that's the last page). We
- * only paginate the daily rollup (PoolDailySnapshot) — at ~365 rows/pool/year
- * a few pages cover years of history for typical pool counts.
- *
- * Failure modes, all designed to fail open rather than blank the dashboard:
- * - Safety cap `SNAPSHOT_MAX_PAGES` reached → return accumulated rows with
- *   `truncated: true, error: null`. Since rows are ordered newest-first, the
- *   missing rows are the oldest ones, so 24h/7d/30d windows stay correct.
- *   No `error` because this is an intentional safety cap, not a fault.
- * - Mid-loop request error AFTER some pages succeeded → keep what we have,
- *   flag `truncated: true` AND set `error` so error-aware consumers (e.g.
- *   the Summary Volume tile) partial-badge correctly. Without the error
- *   signal the degraded state would be invisible to anything that only
- *   checks the error channel.
- * - First-page failure → rethrow. No rows at all means there's nothing to
- *   salvage; the caller surfaces it as `snapshotsAllDailyError` (empty state
- *   with explicit error message, not a misleading `$0` dashboard).
- *
- * Dedup: offset pagination on an append-only table isn't stable under
- * concurrent inserts — a new snapshot at position 0 shifts everything one
- * row right, so the next page's offset overlaps with the previous page's
- * tail. We dedup by `(poolId, timestamp)` (which uniquely identifies a
- * snapshot per the indexer's daily-bucket upsert id) before pushing into the
- * result. This mitigates duplicates; omissions are still theoretically
- * possible but rare and self-heal on the next refresh. A proper fix is
- * keyset pagination — tracked as a follow-up.
- */
-const SNAPSHOT_MAX_PAGES = 100;
-
-// Per-request abort budget. Matches `homepage-og.ts` and `useBridgeGQL`. Without
-// this, a wedged upstream holds the Vercel serverless function open until the
-// platform timeout (10s/60s/300s) and delays TTFB on the SSR path. 5s is more
-// than the p99 of the slowest query measured against Envio's hosted Hasura.
-export const REQUEST_TIMEOUT_MS = 5000;
-
-// Tracks responseKeys that have already triggered a
-// "hasura-snapshot-cap-exhausted" warning so the 30s poll cycle doesn't
-// re-fire the same signal on every refresh. Exported for test-scope
-// `.clear()` so module state doesn't leak across tests.
-/** @internal */
-export const warnedCapKeys = new Set<string>();
-
-// Throttle partial-page failures (mid-pagination exceptions). Without this,
-// a persistently-flaky upstream page fires captureException every 30s poll
-// cycle per chain — quota burn and noise. Keyed by ${network}:${responseKey}
-// so distinct per-chain degradation still surfaces.
-const PARTIAL_PAGE_THROTTLE_MS = 60_000;
-/** @internal */
-export const partialPageLastCapturedAt = new Map<string, number>();
-
-/**
- * Generic paginated fetcher with the Hasura fail-open contract documented
- * above: hard-fail on page 0, preserve and flag truncated on mid-loop failure,
- * log cap exhaustion once per (network, responseKey). `pageSize` must match the
- * GraphQL `limit` passed by `variablesFor`; it is the short-page sentinel.
- *
- * Callers parameterize: variables shape per page, response key, and a dedup
- * key extractor — offset pagination over an append-only table is unstable
- * under concurrent inserts, so dedup is required to keep windowed totals
- * accurate even when a refresh hits mid-write.
- *
- * IMPORTANT: the `limit` a caller passes via `variablesFor` MUST equal
- * `SNAPSHOT_PAGE_SIZE` (1000). That constant is the early-exit sentinel — the
- * loop stops when a page returns fewer than `SNAPSHOT_PAGE_SIZE` rows. A caller
- * passing a smaller `limit` (e.g. 500) would terminate after the first page
- * because `500 < 1000` always holds, silently truncating the result set.
- */
-export async function fetchPaginatedRows<TRow, TVars>(args: {
-  client: GraphQLClient;
-  query: string;
-  responseKey: string;
-  network: string;
-  pageSize: number;
-  variablesFor: (page: number) => TVars;
-  dedupKey: (row: TRow) => string;
-  /** Extra payload merged into the Sentry capture for this responseKey. */
-  extra?: Record<string, unknown>;
-}): Promise<PaginatedPageResult<TRow>> {
-  const {
-    client,
-    query,
-    responseKey,
-    network,
-    pageSize,
-    variablesFor,
-    dedupKey,
-    extra,
-  } = args;
-  const seen = new Set<string>();
-  const rows: TRow[] = [];
-  // Sequential pagination — each iteration breaks early when the page
-  // is short, so we can't batch ahead without risking unnecessary work.
-  for (let page = 0; page < SNAPSHOT_MAX_PAGES; page++) {
-    let batch: TRow[];
-    try {
-      // react-doctor-disable-next-line react-doctor/async-await-in-loop
-      const result = await client.request<Record<string, TRow[]>>({
-        document: query,
-        variables: variablesFor(page) as Record<string, unknown>,
-        signal: AbortSignal.timeout(REQUEST_TIMEOUT_MS),
-      });
-      batch = result[responseKey] ?? [];
-    } catch (err) {
-      if (rows.length === 0) throw err;
-      // Mid-loop failure → preserve fetched pages, flag truncated, throttle
-      // Sentry per (network, responseKey) so the 30s poll can't fan out a storm.
-      const partialKey = `${network}:${responseKey}`;
-      const now = Date.now();
-      const last = partialPageLastCapturedAt.get(partialKey) ?? 0;
-      if (now - last >= PARTIAL_PAGE_THROTTLE_MS) {
-        partialPageLastCapturedAt.set(partialKey, now);
-        Sentry.captureException(err, {
-          tags: {
-            source: "hasura",
-            responseKey,
-            network,
-            degraded: "partial-pages",
-          },
-          extra: { page, rowsFetched: rows.length, ...extra },
-        });
-      }
-      return {
-        rows,
-        truncated: true,
-        error: err instanceof Error ? err : new Error(String(err)),
-      };
-    }
-    for (const row of batch) {
-      const key = dedupKey(row);
-      if (seen.has(key)) continue;
-      seen.add(key);
-      rows.push(row);
-    }
-    if (batch.length < pageSize) {
-      return { rows, truncated: false, error: null };
-    }
-  }
-  // Cap exhaustion: data is genuinely incomplete. One Sentry message per
-  // (network, responseKey) pair — each chain surfaces its own cap event once.
-  const capKey = `${network}:${responseKey}`;
-  if (!warnedCapKeys.has(capKey)) {
-    warnedCapKeys.add(capKey);
-    Sentry.captureMessage("hasura-snapshot-cap-exhausted", {
-      level: "warning",
-      tags: { source: "hasura", responseKey, network },
-      extra: {
-        rowsFetched: rows.length,
-        maxPages: SNAPSHOT_MAX_PAGES,
-        pageSize,
-        ...extra,
-      },
-    });
-  }
-  return { rows, truncated: true, error: null };
-}
-
-// Incremental snapshot cache. Pool×day history is immutable except the current
-// (and, across a midnight boundary, previous) UTC day, so after one successful
-// full-history pagination we only re-fetch the mutable tail and merge.
-// Module-scope on purpose (same lifetime contract as `warnedCapKeys`):
-// client-side it spans the SWR poll loop, server-side it spans ISR
-// regenerations on a warm instance.
-/** @internal Exported for test-scope `.clear()`. */
-export const incrementalRowCache = new Map<
-  string,
-  { variablesKey: string; rows: unknown[]; refreshAfterTimestamp: number }
->();
-
-const poolIdsVariablesKey = (poolIds: readonly string[]) =>
-  [...poolIds].sort().join(",");
-
-type IncrementalSeedArgs = {
-  cacheKey: string;
-  variablesKey: string;
-  rows: unknown[];
-  dedupKey: (row: unknown) => string;
-  timestampOf: (row: unknown) => number;
-  nowMs?: number;
-};
-
-function seedIncrementalRows({
-  cacheKey,
-  variablesKey,
-  rows,
-  dedupKey,
-  timestampOf,
-  nowMs = Date.now(),
-}: IncrementalSeedArgs): void {
-  if (rows.length === 0) return;
-
-  const refreshAfterTimestamp = refreshAfterTimestampForRows(
-    rows,
-    timestampOf,
-    nowMs,
-  );
-  const existing = incrementalRowCache.get(cacheKey);
-  if (existing !== undefined && existing.variablesKey === variablesKey) {
-    const mergedRows = mergeRowsPreservingExisting(
-      existing.rows,
-      rows,
-      dedupKey,
-      timestampOf,
-    );
-    const mergedRefreshAfterTimestamp = refreshAfterTimestampForRows(
-      mergedRows,
-      timestampOf,
-      nowMs,
-    );
-    if (
-      existing.rows.length === mergedRows.length &&
-      existing.refreshAfterTimestamp === mergedRefreshAfterTimestamp
-    ) {
-      return;
-    }
-    incrementalRowCache.set(cacheKey, {
-      variablesKey,
-      rows: mergedRows,
-      refreshAfterTimestamp: mergedRefreshAfterTimestamp,
-    });
-    return;
-  }
-  if (existing !== undefined) return;
-  incrementalRowCache.set(cacheKey, {
-    variablesKey,
-    rows,
-    refreshAfterTimestamp,
-  });
-}
-
-export function seedIncrementalRowCacheFromNetworkData(
-  networkData: readonly NetworkData[],
-): void {
-  for (const data of networkData) {
-    if (
-      data.snapshotsAllDailyError === null &&
-      !data.snapshotsAllDailyTruncated &&
-      data.pools.length > 0
-    ) {
-      seedIncrementalRows({
-        cacheKey: `${data.network.id}:PoolDailySnapshot`,
-        variablesKey: poolIdsVariablesKey(data.pools.map((pool) => pool.id)),
-        rows: data.snapshotsAllDaily,
-        dedupKey: (row) => {
-          const snapshot = row as PoolSnapshotWindow;
-          return `${snapshot.poolId}-${snapshot.timestamp}`;
-        },
-        timestampOf: (row) => Number((row as PoolSnapshotWindow).timestamp),
-      });
-    }
-
-    // Intentionally do not seed `PoolDailyFeeSnapshot`: older fee rows can be
-    // healed later when token metadata resolves, so fee history must keep using
-    // full pagination until the source exposes a version/invalidation cursor.
-  }
-}
-
-const SECONDS_PER_DAY = 86_400;
-
-// Oldest timestamp that can still mutate: today's UTC day bucket, minus one
-// extra day so a poll that crosses midnight still re-fetches the final updates
-// to yesterday's row.
-function mutableDayCutoff(nowMs: number): number {
-  const todayMidnightUtc =
-    Math.floor(nowMs / 1000 / SECONDS_PER_DAY) * SECONDS_PER_DAY;
-  return todayMidnightUtc - SECONDS_PER_DAY;
-}
-
-function refreshAfterTimestampForRows<TRow>(
-  rows: readonly TRow[],
-  timestampOf: (row: TRow) => number,
-  nowMs: number,
-): number {
-  const newestIndexedTimestamp = rows.reduce(
-    (newest, row) => Math.max(newest, timestampOf(row)),
-    0,
-  );
-  return Math.min(mutableDayCutoff(nowMs), newestIndexedTimestamp);
-}
-
-function mergeRowsPreservingExisting<TRow>(
-  existingRows: readonly TRow[],
-  incomingRows: readonly TRow[],
-  dedupKey: (row: TRow) => string,
-  timestampOf: (row: TRow) => number,
-): TRow[] {
-  const merged = new Map<string, TRow>();
-  for (const row of existingRows) merged.set(dedupKey(row), row);
-  for (const row of incomingRows) {
-    const key = dedupKey(row);
-    if (!merged.has(key)) merged.set(key, row);
-  }
-  return [...merged.values()].sort(
-    (a, b) =>
-      timestampOf(b) - timestampOf(a) || dedupKey(b).localeCompare(dedupKey(a)),
-  );
-}
-
-function shouldCacheCompleteRows<TRow>(
-  result: PaginatedPageResult<TRow>,
-): boolean {
-  return !result.truncated && result.error === null && result.rows.length > 0;
-}
-
-function surfacedIncrementalError(
-  error: Error | null,
-  surfaceIncrementalError: boolean,
-): Error | null {
-  return surfaceIncrementalError ? error : null;
-}
-
-async function fetchPaginatedRowsIncremental<TRow, TVars>(args: {
-  client: GraphQLClient;
-  query: string;
-  responseKey: string;
-  network: string;
-  pageSize: number;
-  /** Serialized non-timestamp variables; a change invalidates the cache. */
-  variablesKey: string;
-  variablesFor: (page: number, afterTimestamp: number) => TVars;
-  dedupKey: (row: TRow) => string;
-  timestampOf: (row: TRow) => number;
-  nowMs?: number;
-  surfaceIncrementalError?: boolean;
-  extra?: Record<string, unknown>;
-}): Promise<PaginatedPageResult<TRow>> {
-  const {
-    client,
-    query,
-    responseKey,
-    network,
-    pageSize,
-    variablesKey,
-    variablesFor,
-    dedupKey,
-    timestampOf,
-    nowMs = Date.now(),
-    surfaceIncrementalError = true,
-    extra,
-  } = args;
-  const cacheKey = `${network}:${responseKey}`;
-  const cached = incrementalRowCache.get(cacheKey);
-  const paginate = (afterTimestamp: number) =>
-    fetchPaginatedRows<TRow, TVars>({
-      client,
-      query,
-      responseKey,
-      network,
-      pageSize,
-      variablesFor: (page) => variablesFor(page, afterTimestamp),
-      dedupKey,
-      ...(extra !== undefined ? { extra } : {}),
-    });
-
-  if (cached === undefined || cached.variablesKey !== variablesKey) {
-    const full = await paginate(0);
-    // Only complete results seed the cache — truncated/errored full fetches
-    // retry the full pagination next cycle, identical to legacy behavior.
-    // Empty complete histories are also left uncached so a backfilling indexer
-    // does not skip newly indexed catch-up days on the next poll.
-    if (shouldCacheCompleteRows(full)) {
-      incrementalRowCache.set(cacheKey, {
-        variablesKey,
-        rows: full.rows,
-        refreshAfterTimestamp: refreshAfterTimestampForRows(
-          full.rows,
-          timestampOf,
-          nowMs,
-        ),
-      });
-    }
-    return full;
-  }
-
-  const cachedRows = cached.rows as TRow[];
-  let tail: PaginatedPageResult<TRow>;
-  try {
-    tail = await paginate(cached.refreshAfterTimestamp);
-  } catch (err) {
-    // First-page failure on the incremental path: degrade to cached history
-    // (mirrors the mid-loop fail-open contract on SNAPSHOT_PAGE_SIZE).
-    const error = err instanceof Error ? err : new Error(String(err));
-    return {
-      rows: cachedRows,
-      truncated: true,
-      error: surfacedIncrementalError(error, surfaceIncrementalError),
-      mutableTailError: error,
-    };
-  }
-
-  const merged = new Map<string, TRow>();
-  for (const row of cachedRows) merged.set(dedupKey(row), row);
-  for (const row of tail.rows) merged.set(dedupKey(row), row);
-  // Keep the legacy newest-first ordering — `fetchNetworkData` reads the LAST
-  // element as the oldest fetched row (`oldestFetchedTs`).
-  const rows = [...merged.values()].sort(
-    (a, b) =>
-      timestampOf(b) - timestampOf(a) || dedupKey(b).localeCompare(dedupKey(a)),
-  );
-
-  if (shouldCacheCompleteRows(tail)) {
-    incrementalRowCache.set(cacheKey, {
-      variablesKey,
-      rows,
-      refreshAfterTimestamp: refreshAfterTimestampForRows(
-        rows,
-        timestampOf,
-        nowMs,
-      ),
-    });
-  }
-
-  return {
-    rows,
-    truncated: tail.truncated,
-    error: surfacedIncrementalError(tail.error, surfaceIncrementalError),
-    mutableTailError:
-      tail.error ??
-      (tail.truncated
-        ? new Error("Snapshot incremental tail pagination truncated")
-        : null),
-  };
-}
-
-/**
- * Daily rollup pagination — ~365 rows/pool/year, so for typical history a
- * few pages cover everything.
- */
-export async function fetchAllDailySnapshotPages(
-  client: GraphQLClient,
-  poolIds: string[],
-  network: string,
-  nowMs?: number,
-): Promise<SnapshotPageResult> {
-  return fetchPaginatedRowsIncremental<PoolSnapshotWindow, unknown>({
-    client,
-    query: POOL_DAILY_SNAPSHOTS_ALL,
-    responseKey: "PoolDailySnapshot",
-    network,
-    pageSize: SNAPSHOT_PAGE_SIZE,
-    variablesKey: poolIdsVariablesKey(poolIds),
-    variablesFor: (page, afterTimestamp) => ({
-      poolIds,
-      afterTimestamp,
-      limit: SNAPSHOT_PAGE_SIZE,
-      offset: page * SNAPSHOT_PAGE_SIZE,
-    }),
-    dedupKey: (s) => `${s.poolId}-${s.timestamp}`,
-    timestampOf: (s) => Number(s.timestamp),
-    ...(nowMs !== undefined ? { nowMs } : {}),
-    extra: { poolCount: poolIds.length },
-  });
-}
-
-/**
- * Paginate `BrokerDailySnapshot` rows for a chain — already filtered to
- * `routedViaV3Router=false` server-side. Row count grows as
- * `(daily_distinct_providers × days)`, so a single chain stays under the
- * 1000-row Hasura cap for years; pagination is here for safety.
- */
-async function fetchAllBrokerDailySnapshotPages(
-  client: GraphQLClient,
-  chainId: number,
-  network: string,
-): Promise<PaginatedPageResult<BrokerDailySnapshotRow>> {
-  return fetchPaginatedRows<BrokerDailySnapshotRow, unknown>({
-    client,
-    query: BROKER_DAILY_SNAPSHOTS_ALL,
-    responseKey: "BrokerDailySnapshot",
-    network,
-    pageSize: SNAPSHOT_PAGE_SIZE,
-    variablesFor: (page) => ({
-      chainId,
-      limit: SNAPSHOT_PAGE_SIZE,
-      offset: page * SNAPSHOT_PAGE_SIZE,
-    }),
-    // The schema id `{chainId}-{provider}-{router|direct}-{day}` is the
-    // canonical row key — required because two providers can share the same
-    // (timestamp, volumeUsdWei, swapCount) tuple at low traffic.
-    dedupKey: (row) => row.id,
-  });
-}
-
-/**
- * Paginate `LiquidityPosition` rows for a set of pool IDs, deduplicating
- * by lowercase address. One LP may hold positions in multiple pools, so
- * the same wallet address can appear across rows — dedup collapses them
- * to a unique-address list. Same fail-open semantics as the daily snapshot
- * path: hard-fail on page 0, preserve and flag truncated on mid-loop failure.
- */
-async function fetchAllLpAddressPages(
-  client: GraphQLClient,
-  poolIds: string[],
-  network: string,
-): Promise<PaginatedPageResult<{ address: string }>> {
-  return fetchPaginatedRows<{ address: string }, unknown>({
-    client,
-    query: UNIQUE_LP_ADDRESSES,
-    responseKey: "LiquidityPosition",
-    network,
-    pageSize: SNAPSHOT_PAGE_SIZE,
-    variablesFor: (page) => ({
-      poolIds,
-      limit: SNAPSHOT_PAGE_SIZE,
-      offset: page * SNAPSHOT_PAGE_SIZE,
-    }),
-    dedupKey: (r) => r.address.toLowerCase(),
-    extra: { poolCount: poolIds.length },
-  });
-}
-
-/**
- * Per-pool fee-snapshot pagination — pool×day cardinality (~30 pools × ~430
- * days) easily exceeds the 1000-row cap, so we paginate. Same fail-open
- * semantics as the daily snapshot path. Dedup keyed off `id` since
- * `PoolDailyFeeSnapshot.id = "{chainId}-{poolAddress}-{dayTs}"` is unique by
- * construction. Exported because the slim `useProtocolFees` hook fetches its
- * own snapshots — the homepage path in `fetchNetworkData` doesn't render the
- * per-pool revenue table so it leaves `feeSnapshots` empty.
- */
-export async function fetchAllFeeSnapshotPages(
-  client: GraphQLClient,
-  chainId: number,
-  network: string,
-): Promise<PaginatedPageResult<PoolDailyFeeSnapshot>> {
-  return fetchPaginatedRows<PoolDailyFeeSnapshot, unknown>({
-    client,
-    query: POOL_DAILY_FEE_SNAPSHOTS_PAGE,
-    responseKey: "PoolDailyFeeSnapshot",
-    network,
-    pageSize: SNAPSHOT_PAGE_SIZE,
-    variablesFor: (page) => ({
-      chainId,
-      afterTimestamp: 0,
-      limit: SNAPSHOT_PAGE_SIZE,
-      offset: page * SNAPSHOT_PAGE_SIZE,
-    }),
-    dedupKey: (s) => s.id,
-  });
-}
 
 /** @internal Exported for testing only. */
 // eslint-disable-next-line complexity, max-lines-per-function, sonarjs/cognitive-complexity -- Existing network orchestrator owns many fail-open slices; keep the exception visible in source instead of carrying a churn-prone baseline tuple.
